python_scripts/producer/requirements.txt
retry
openlineage-python
confluent-kafka
